src/cloudbot/requirements.txt
wheel
google-cloud-logging
google-cloud-secret-manager
flask
PyGithub
numpy
#octokit
google-generativeai